requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.0
fonttools==4.40.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.11.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchsummary==1.5.1
torchvision==0.15.2+cu117
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
zipp==3.15.0
